requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.1
pillow>=10.0.0
mss>=9.0.0
pyautogui>=0.9.54
sounddevice>=0.4.6
numpy>=1.26.0
pyperclip>=1.8.0
uiautomation>=2.0.18
